+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 0.9.5
+- Reverse debugging: Conditional breakpoints.
 - Fixed moving of PC to cursor during reverse debugging.
 - Changed reverse debug decoration to blue background only due to performance reasons.